\name{GEEaSPU-package}
\alias{GEEaSPU-package}
\docType{package}
\title{
Adaptive Association Tests for Multiple Phenotypes using Generalized Estimating Equations (GEE)
}
\description{
Provides adaptive association tests for SNP level, gene level and pathway level analyses.
}
\details{
Functions for adaptive association tests including GEEaSPU, 
GEEaSPUset and GEEaSPUpath. These methods use a weighting scheme for 
testing associations with multiple phenotypes. GEEaSPU function computes a SNP
level p value; GEEaSPUset function can be used for the SNP-set (gene) level
association test, while GEEaSPUpath function is for the gene-set (pathway) level analysis.
}
\author{
Junghi Kim and Wei Pan
}

\references{
Kim J, Zhang Y, and Pan W (2016). "Powerful and adaptive testing for multi-trait and multi-SNP associations with GWAS and sequencing data." \emph{Genetics} 203(2), 715--731.


Zhang Y, Xu Z, Shen, X and  Pan W (2014). "Testing for association with multiple traits in generalized estimation equations, with application to neuroimaging data." \emph{NeuroImage} 96, 309--325.
}
